changed generator data types

+++ building_road_segmentation/data_generator.py
@@ -7,6 +7,7 @@ import tensorflow.keras.utils
 import numpy as np
 import PIL
 
+
 class read_data(tf.keras.utils.Sequence):
 
     def __init__(self, x_set, y_set, batch_size):
@@ -17,14 +18,14 @@ class read_data(tf.keras.utils.Sequence):
         return math.ceil(len(self.x) / self.batch_size)
 
     def __getitem__(self, idx):
-        batch_x = self.x[idx * self.batch_size:(idx + 1) *
-        self.batch_size]
-        batch_y = self.y[idx * self.batch_size:(idx + 1) *
-        self.batch_size]
+        batch_x = self.x[idx * self.batch_size:(idx + 1) * self.batch_size]
+        batch_y = self.y[idx * self.batch_size:(idx + 1) * self.batch_size]
 
-        return np.array([np.array(PIL.Image.open(im)) for im in batch_x], dtype=np.float32), np.array([
-            np.load(file_name)
-               for file_name in batch_y])
+        return np.array([np.array(PIL.Image.open(im)) for im in batch_x],
+                        dtype=np.float32), np.array([
+                            np.load(file_name).astype(np.float32)
+                            for file_name in batch_y
+                        ])
 
 
 class test_data(tf.keras.utils.Sequence):
@@ -37,9 +38,7 @@ class test_data(tf.keras.utils.Sequence):
         return math.ceil(len(self.x) / self.batch_size)
 
     def __getitem__(self, idx):
-        batch_x = self.x[idx * self.batch_size:(idx + 1) *
-        self.batch_size]
-        batch_y = self.y[idx * self.batch_size:(idx + 1) *
-        self.batch_size]
+        batch_x = self.x[idx * self.batch_size:(idx + 1) * self.batch_size]
+        batch_y = self.y[idx * self.batch_size:(idx + 1) * self.batch_size]
 
         return batch_x, batch_y
